// deleghebowcl/src/app/model/tokenResponse.ts
/**
 * DELEGHEBO_API
 * API del back office di Deleghe.
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TokenResponse {

  codice_azienda?: string;
  codice_fiscale?: string;
  esito?: string;


}